create metadata service class
